+++ src/components/index.js
@@ -1,9 +1,8 @@
 export { default as Button } from './Button/Button';
-export { default as Alert } from './Alert/Alert';
 export { default as Input } from './Input/Input';
 export { default as Select } from './Select/Select';
 export { default as SelectMobile } from './SelectMobile/SelectMobile';
-// export { default as Spinner } from './Spinner/Spinner';
+export { default as Text } from './Text/Text';
 
 // Legacy components