+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.5] - 20/04/2022
+## [0.4.0] - 06/05/2022
+### Add
+- Add new component: Text
+
+## [0.3.5] - 05/05/2022
 ### fixed
 - Fonts Poppins
 - Fix error SelectMobile component